Commit first version of code
This commit includes the first version of the code. This still needs
some improvements, but I was able to use this version to successfully
complete the Teams automation for GT6000.

At some point, I wonder if it might make more sense to record the
coordinates in a different format and then have a separate utility that
is able to play an "automouse script" maybe, rather than recording the
coordinates directly in a bash / shell script (that the user can then
modify for whatever application).